Improve API, cleanup splat naming
