+++ mujoco_warp/__init__.py
@@ -48,7 +48,6 @@ from mujoco_warp._src.forward import implicit as implicit
 from mujoco_warp._src.forward import rungekutta4 as rungekutta4
 from mujoco_warp._src.forward import step1 as step1
 from mujoco_warp._src.forward import step2 as step2
-from mujoco_warp._src.gaussian import load_gaussian_ply as load_gaussian_ply
 from mujoco_warp._src.history import init_ctrl_history as init_ctrl_history
 from mujoco_warp._src.history import init_sensor_history as init_sensor_history
 from mujoco_warp._src.history import read_ctrl as read_ctrl
