+++ mujoco_warp/_src/types.py
@@ -2491,15 +2491,14 @@ class RenderContext:
       fallback are controlled by `use_ambient_lighting`.
     geom_ray_types: tuple of GeomType int values present in the scene, used to
       statically eliminate unused intersection branches in the ray-cast kernels.
-    gaussian_transforms: Gaussian transforms
-    gaussian_scales: Gaussian scales
-    gaussian_rgba: Gaussian colors and opacities
-    gaussian_bvh: Gaussian BVH
-    gaussian_bvh_id: Gaussian BVH id
-    gaussian_lower: Gaussian lower bounds
-    gaussian_upper: Gaussian upper bounds
-    gaussian_min_response: minimum rendered Gaussian response
-    gaussian_count: number of Gaussians
+    splat_transforms: Splat transforms
+    splat_scales: Splat scales
+    splat_rgba: Splat colors and opacities
+    splat_bvh: Splat BVH
+    splat_bvh_id: Splat BVH id
+    splat_lower: Splat lower bounds
+    splat_upper: Splat upper bounds
+    splat_count: Number of splats
   """
 
   nrender: int
@@ -2564,13 +2563,12 @@ class RenderContext:
   enable_per_light_ambient: bool
   light_attenuation_is_default: bool
   has_spot_lights: bool
-  gaussian_transforms: array("*", wp.transform)
-  gaussian_scales: array("*", wp.vec3)
-  gaussian_rgba: array("*", wp.vec4)
-  gaussian_bvh: Optional[wp.Bvh]
-  gaussian_bvh_id: wp.uint64
-  gaussian_lower: array("*", wp.vec3)
-  gaussian_upper: array("*", wp.vec3)
-  gaussian_min_response: float
-  gaussian_count: int
+  splat_transforms: array("*", wp.transform)
+  splat_scales: array("*", wp.vec3)
+  splat_rgba: array("*", wp.vec4)
+  splat_bvh: Optional[wp.Bvh]
+  splat_bvh_id: wp.uint64
+  splat_lower: array("*", wp.vec3)
+  splat_upper: array("*", wp.vec3)
+  splat_count: int
   geom_ray_types: tuple = ()
